+++ src/main/java/com/codesquad/airbnb/controller/UserController.java
@@ -1,8 +1,5 @@
 package com.codesquad.airbnb.controller;
 
-import com.codesquad.airbnb.GoogleUser;
-import com.codesquad.airbnb.OAuthToken;
-import com.codesquad.airbnb.domain.User;
 import com.codesquad.airbnb.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,21 +23,7 @@ public class UserController {
 
     @GetMapping("/oauth/google/callback") // user가 정보사용 승인하면 google이 String code를 queryString으로 보내 준다(?code="")
     public ResponseEntity oauthLogin(String code) {
-        ResponseEntity<String> accessTokenResponse = userService.createPost(code);
-        // response안에 있는 정보를 OAuth 액세스 토큰으로 변환하기
-        OAuthToken oAuthToken = userService.getAccessToken(accessTokenResponse);
-        logger.info("Access Token: {}", oAuthToken.getAccessToken());
-
-        ResponseEntity<String> userInfoResponse = userService.createGet(oAuthToken);
-        GoogleUser googleUser = userService.getUserInfo(userInfoResponse);
-        logger.info("Google User Name: {}", googleUser.getName());
-        
-        if (!userService.isJoinedUser(googleUser)) {
-            User user = googleUser.toUser(oAuthToken.getAccessToken());
-            userService.save(user); // 회원가입
-        }
-        // 로그인
-
+        userService.oauthLogin(code);
         return new ResponseEntity("로그인 성공", HttpStatus.OK);
     }
 }
